refactor(client-runtime): share realtime voice events
